04/08/2020, 20:00
- Test Engagement

- Test Loan B

06/08/2020, 19:00
- Test download cv, response base 64


- Test engagement

07/08/2020, 19:00

1. Engagement list (done)
2. Loan A List (done0)
3. Approve / Reject Done
4. show status loan Done

10/08/2020, 22:15
1. update case-id string format
2. update loan-b id string format

14/08/2020
1. update response pdf using jsonobject<base64>

19/08/2020 14:55
1. update response pdf using jsonobject<base64>, method downloadCV


19/08/2020 20:00
1. create feature disbursement(progress)

24/08/2020 19:00
1. create feature disbursement(progress)
2. create feature reimbursement(progress)

25/08/2020 19:00
1. create feature disbursement(progress)
2. create feature reimbursement(progress)
2. create feature case document(progress)

01/09/2020 17:18
1. create feature disbursement(progress)
2. create feature reimbursement(progress)
2. create feature case document(progress)

02/09/2020 19:00
1. create feature disbursement(progress)
2. create feature reimbursement(progress)
2. create feature case document(progress)

04/08/2020, 20:00

- Test Engagement

- Test Loan B

06/08/2020, 19:00
- Test download cv, response base 64


- Test engagement

07/08/2020, 19:00

1. Engagement list (done)
2. Loan A List (done0)
3. Approve / Reject Done
4. show status loan Done

10/08/2020, 22:15
1. update case-id string format
2. update loan-b id string format

14/08/2020
1. update response pdf using jsonobject<base64>

19/08/2020 14:55
1. update response pdf using jsonobject<base64>, method downloadCV


19/08/2020 20:00
1. create feature disbursement(progress)

24/08/2020 19:00
1. create feature disbursement(progress)
2. create feature reimbursement(progress)

25/08/2020 19:00
1. create feature disbursement(progress)
2. create feature reimbursement(progress)
2. create feature case document(progress)

01/09/2020 17:18
1. create feature disbursement(progress)
2. create feature reimbursement(progress)
2. create feature case document(progress)

02/09/2020 20:00
1. create feature disbursement(progress)
2. create feature reimbursement(progress)
2. create feature case document(progress)

08/09/2020 19:00
1. create feature disbursement(progress)
2. create feature reimbursement(progress)
2. create feature case document(progress)

10/09/2020 22:00
1. create feature disbursement(progress)
2. create feature reimbursement(progress)
2. create feature case document(progress)

20/09/2020 12:00
1.
/employee/managed-employee/in-active (testing)
/employee/managed-employee/active (testing)
/employee/managed-employee/resign (testing)
2.
/disburse/{id_loan}/loan-a (testing)
/disburse/{id_loan}/loan-b (testing)
/disbursements/loans (testing)
/disbursements/loan-a  (testing)
/disbursements/loan-b  (testing)
/disbursement/{id_loan} (testing)

3.
/oust-standing/loan-a (Progress)
/oust-standing/loan-a/employee-id (Progress)
/oust-standing/loan-b (Progress)
/oust-standing/loan-b/case-id (Progress)


24/09/2020 19:00
1.
/employee/managed-employee/in-active (testing)
/employee/managed-employee/active (testing)
/employee/managed-employee/resign (testing)
2.
/disburse/{id_loan}/loan-a (testing)
/disburse/{id_loan}/loan-b (testing)
/disbursements/loans (testing)
/disbursements/loan-a  (testing)
/disbursements/loan-b  (testing)
/disbursement/{id_loan} (testing)

3.
/oust-standing/loan-a (Progress)
/oust-standing/loan-a/employee-id (Progress)
/oust-standing/loan-b (Progress)
/oust-standing/loan-b/case-id (Progress)

29/09/2020 19:00
1.
/employee/managed-employee/in-active (testing)
/employee/managed-employee/active (testing)
/employee/managed-employee/resign (testing)
2.
/disburse/{id_loan}/loan-a (testing)
/disburse/{id_loan}/loan-b (testing)
/disbursements/loans (testing)
/disbursements/loan-a  (testing)
/disbursements/loan-b  (testing)
/disbursement/{id_loan} (testing)

3.
/oust-standing/loan-a (Progress)
/oust-standing/loan-a/employee-id (Progress)
/oust-standing/loan-b (Progress)
/oust-standing/loan-b/case-id (Progress)

06/10/2020 22:00
1.
/employee/managed-employee/in-active (done)
/employee/managed-employee/resign (done)
/employee/managed-employee/update-profile (testing UI)
/employee/managed-employee/account (testing UI)
/employee/approval/account (testing UI)
2.
/disburse/{id_loan}/loan-a (testing UI)
/disburse/{id_loan}/loan-b (testing UI)
/disbursements/loans (testing UI)
/disbursements/loan-a  (testing UI)
/disbursements/loan-b  (testing UI)
/disbursement/{id_loan} (testing UI)

3.
/oust-standing/loan-a (Progress)
/oust-standing/loan-a/employee-id (Progress)
/oust-standing/loan-b (Progress)
/oust-standing/loan-b/case-id (Progress)

08/10/2020 19:30
1.
/employee/managed-employee/in-active (done)
/employee/managed-employee/resign (done)
/employee/managed-employee/update-profile (testing UI)
/employee/managed-employee/account (testing UI)
/employee/approval/account (testing UI)
2.
/disburse/{id_loan}/loan-a (testing UI)
/disburse/{id_loan}/loan-b (testing UI)
/disbursements/loans (testing UI)
/disbursements/loan-a  (testing UI)
/disbursements/loan-b  (testing UI)
/disbursement/{id_loan} (testing UI)

3.
/oust-standing/loan-a (Progress)
/oust-standing/loan-a/employee-id (Progress)
/oust-standing/loan-b (Progress)
/oust-standing/loan-b/case-id (Progress)

+++ src/main/java/com/lawfirm/apps/repo/interfaces/OutStandingLoanARepoIface.java
@@ -35,5 +35,5 @@ public interface OutStandingLoanARepoIface {
 //    List<OutStandingLoanB> findByCaseId(Long paramLong);
     OutStandingLoanA findByCaseId(String param);
 
-    OutStandingLoanA findBy(Long idEmployee, String taxyear, Integer disburseId);
+    OutStandingLoanA findBy(Long idEmployee, String taxyear, Long disburseId);
 }
